🧹 Chore: cleanup tests
